Created separate file for ebola test data
